Use System.Text.Json source generator
